Library updated to 0.0.4-SNAPSHOT Removed all test from libraries Reimplemented dependencies of ImageView implementations: * library module now is a main dependency for other ImageView implementations
Implemented picasso library
Removed bad setElevation implementation
