+++ app/src/main/java/com/imageview/sample/MainActivity.java
@@ -29,7 +29,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 
-import com.imageview.ImageView;
+import com.imageview.glide.ImageView;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -39,6 +39,7 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         final ImageView imageView = (ImageView) findViewById(R.id.image_view);
+        //imageView.setClipToOutline(true);
         imageView.setImageURL("https://avatars2.githubusercontent.com/u/8938207?v=3&s=460");
 
         Button button = (Button) findViewById(R.id.button_test);
