+++ gradle.properties
@@ -17,7 +17,7 @@
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 #-SNAPSHOT
-version=0.0.3-SNAPSHOT
+version=0.0.4-SNAPSHOT
 group=com.github.vihtarb
 
 POM_DESCRIPTION=A extended image view for android
