+++ library-glide/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="com.imageview">
+<manifest package="com.imageview.glide">
 
     <application/>
