+++ library-picasso/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=ImageView
+POM_ARTIFACT_ID=imageview-picasso
+POM_PACKAGING=aar
