+++ library-picasso/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest package="com.imageview.picasso">
+
+    <application/>
+
+</manifest>
