+++ library-picasso/src/main/res/values/attrs.xml
